requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
pandas>=2.0.0
openpyxl>=3.0.0
Flask-CORS==4.0.0